requirements.txt
anyio
asgiref
attrs
autobahn
Automat
boto3
certifi
cffi
channels
channels_redis
charset-normalizer
click
colorama
constantly
cryptography
daphne
Django
django-axes
django-cors-headers
django-csp
django-pwa
django-storages
dj-database-url
h11
httptools
hyperlink
idna
incremental
lxml
msgpack
oauthlib
packaging
pillow
pyasn1
pyasn1_modules
pycparser
PyMuPDF
pyOpenSSL
python-pptx
python-dotenv
PyYAML
redis
requests
requests-oauthlib
service-identity
setuptools
six
sniffio
sqlparse
Twisted
txaio
typing_extensions
tzdata
ua-parser
ua-parser-builtins
urllib3
user-agents
uvicorn
watchfiles
websockets
whitenoise
xlsxwriter
zope.interface
